Move DeedManager to deed.ts
